Modulo6/Practice/requirements.txt
# Manipulación de datos
numpy==1.24.4
pandas==2.1.1

# Visualización de datos
matplotlib==3.8.0
seaborn==0.12.2

# Reducción de dimensionalidad
scikit-learn==1.3.1
umap-learn==0.5.4

# PROYECTO BACKEND
flask==3.0.0

# Guardar o cargar pesos del modelo
joblib==1.3.2
